+++ test/steps/v2_steps.py
@@ -295,6 +295,13 @@ def parse_ledger(context, tot, online, roundNum):
     assert context.response["current_round"] == int(roundNum)
 
 
+@when(
+    'we make an Account Information call against account "{account}" with exclude "{exclude:MaybeString}"'
+)
+def acc_info(context, account, exclude):
+    context.response = context.acl.account_info(account, exclude=exclude)
+
+
 @when('we make an Account Information call against account "{account}"')
 def acc_info(context, account):
     context.response = context.acl.account_info(account)
@@ -312,6 +319,22 @@ def acc_info_any(context):
 )
 def parse_acc_info(context, address):
     assert context.response["address"] == address
+
+
+@when(
+    'we make an Account Asset Information call against account "{account}" assetID {assetID}'
+)
+def acc_asset_info(context, account, assetID):
+    context.response = context.acl.account_asset_info(account, assetID)
+
+
+@when(
+    'we make an Account Application Information call against account "{account}" applicationID {applicationID}'
+)
+def acc_application_info(context, account, applicationID):
+    context.response = context.acl.account_application_info(
+        account, applicationID
+    )
 
 
 @when("we make a GetAssetByID call for assetID {asset_id}")
@@ -370,14 +393,13 @@ def check_asset_balance(context, numaccounts, account, isfrozen, amount):
 
 
 @when(
-    'we make a Lookup Asset Balances call against asset index {index} with limit {limit} afterAddress "{afterAddress:MaybeString}" round {block} currencyGreaterThan {currencyGreaterThan} currencyLessThan {currencyLessThan}'
+    'we make a Lookup Asset Balances call against asset index {index} with limit {limit} afterAddress "{afterAddress:MaybeString}" currencyGreaterThan {currencyGreaterThan} currencyLessThan {currencyLessThan}'
 )
 def asset_balance(
     context,
     index,
     limit,
     afterAddress,
-    block,
     currencyGreaterThan,
     currencyLessThan,
 ):
@@ -387,7 +409,6 @@ def asset_balance(
         next_page=None,
         min_balance=int(currencyGreaterThan),
         max_balance=int(currencyLessThan),
-        block=int(block),
     )
 
 
@@ -408,6 +429,64 @@ def parse_asset_balance(
     assert context.response["balances"][int(idx)]["amount"] == int(amount)
     assert context.response["balances"][int(idx)]["is-frozen"] == (
         frozenState == "true"
+    )
+
+
+@when(
+    'we make a LookupAccountAssets call with accountID "{account}" assetID {asset_id} includeAll "{includeAll:MaybeBool}" limit {limit} next "{next:MaybeString}"'
+)
+def lookup_account_assets(context, account, asset_id, includeAll, limit, next):
+    context.response = context.icl.lookup_account_assets(
+        account,
+        asset_id=int(asset_id),
+        include_all=includeAll,
+        limit=int(limit),
+        next_page=next,
+    )
+
+
+@when(
+    'we make a LookupAccountCreatedAssets call with accountID "{account}" assetID {asset_id} includeAll "{includeAll:MaybeBool}" limit {limit} next "{next:MaybeString}"'
+)
+def lookup_account_created_assets(
+    context, account, asset_id, includeAll, limit, next
+):
+    context.response = context.icl.lookup_account_asset_by_creator(
+        account,
+        asset_id=int(asset_id),
+        include_all=includeAll,
+        limit=int(limit),
+        next_page=next,
+    )
+
+
+@when(
+    'we make a LookupAccountAppLocalStates call with accountID "{account}" applicationID {application_id} includeAll "{includeAll:MaybeBool}" limit {limit} next "{next:MaybeString}"'
+)
+def lookup_account_applications(
+    context, account, application_id, includeAll, limit, next
+):
+    context.response = context.icl.lookup_account_application_local_state(
+        account,
+        application_id=int(application_id),
+        include_all=includeAll,
+        limit=int(limit),
+        next_page=next,
+    )
+
+
+@when(
+    'we make a LookupAccountCreatedApplications call with accountID "{account}" applicationID {application_id} includeAll "{includeAll:MaybeBool}" limit {limit} next "{next:MaybeString}"'
+)
+def lookup_account_created_applications(
+    context, account, application_id, includeAll, limit, next
+):
+    context.response = context.icl.lookup_account_application_by_creator(
+        account,
+        application_id=int(application_id),
+        include_all=includeAll,
+        limit=int(limit),
+        next_page=next,
     )
 
 
@@ -791,6 +870,13 @@ def lookup_account(context, account, block):
     context.response = context.icl.account_info(account, int(block))
 
 
+@when(
+    'we make a Lookup Account by ID call against account "{account}" with exclude "{exclude:MaybeString}"'
+)
+def lookup_account(context, account, exclude):
+    context.response = context.icl.account_info(account, exclude=exclude)
+
+
 @when("we make any LookupAccountByID call")
 def lookup_account_any(context):
     context.response = context.icl.account_info(
@@ -897,6 +983,11 @@ def search_application(context, app_id):
     context.response = context.icl.search_applications(int(app_id))
 
 
+@when('we make a SearchForApplications call with creator "{creator}"')
+def search_application(context, creator):
+    context.response = context.icl.search_applications(creator=creator)
+
+
 @when(
     "we make a Search Accounts call with assetID {index} limit {limit} currencyGreaterThan {currencyGreaterThan} currencyLessThan {currencyLessThan} and round {block}"
 )
@@ -936,6 +1027,14 @@ def search_accounts(
         block=int(block),
         auth_addr=authAddr,
     )
+
+
+@when('we make a Search Accounts call with exclude "{exclude:MaybeString}"')
+def search_accounts(
+    context,
+    exclude,
+):
+    context.response = context.icl.accounts(exclude=exclude)
 
 
 @when(
@@ -2985,7 +3084,7 @@ def deserialize_json_to_contract(context):
 
 
 @given(
-    'a dryrun response file "{dryrun_response_file}" and a transaction id "{txn_id}"'
+    'a dryrun response file "{dryrun_response_file}" and a transaction at index "{txn_id}"'
 )
 def parse_dryrun_response_object(context, dryrun_response_file, txn_id):
     dir_path = os.path.dirname(os.path.realpath(__file__))
@@ -3001,21 +3100,20 @@ def parse_dryrun_response_object(context, dryrun_response_file, txn_id):
     ]
 
 
-@when("I call app trace")
-def dryrun_app_trace(context):
-    context.dryrun_trace = context.dryrun_txn_result.app_trace()
-
-
-@then('the output should equal "{golden_file}"')
-def dryrun_compare_golden(context, golden_file):
+@then('calling app trace produces "{app_trace_file}"')
+def dryrun_compare_golden(context, app_trace_file):
     dir_path = os.path.dirname(os.path.realpath(__file__))
     dir_path = os.path.dirname(os.path.dirname(dir_path))
-    with open(dir_path + "/test/features/resources/" + golden_file, "r") as f:
-        golden_expected = f.read()
+    with open(
+        dir_path + "/test/features/resources/" + app_trace_file, "r"
+    ) as f:
+        trace_expected = f.read()
 
-    assert (
-        golden_expected == context.dryrun_trace
-    ), "Expected '{}' got '{}'".format(golden_expected, context.dryrun_trace)
+    dryrun_trace = context.dryrun_txn_result.app_trace()
+
+    assert trace_expected == dryrun_trace, "Expected \n{}\n got \n{}\n".format(
+        trace_expected, dryrun_trace
+    )
 
 
 @then(